requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.14.0